separating reputation models
