+++ src/reputation/related_models/paid_status_mixin.py
@@ -1,0 +1,34 @@
+from django.db import models
+from django.utils import timezone
+
+
+class PaidStatusModelMixin(models.Model):
+    FAILED = "FAILED"
+    PAID = "PAID"
+    PENDING = "PENDING"
+    PAID_STATUS_CHOICES = [
+        (FAILED, FAILED),
+        (PAID, PAID),
+        (PENDING, PENDING),
+    ]
+
+    class Meta:
+        abstract = True
+
+    paid_date = models.DateTimeField(default=None, null=True)
+    paid_status = models.CharField(
+        max_length=255, choices=PAID_STATUS_CHOICES, default=None, null=True
+    )
+
+    def set_paid_failed(self):
+        self.paid_status = self.FAILED
+        self.save()
+
+    def set_paid_pending(self):
+        self.paid_status = self.PENDING
+        self.save()
+
+    def set_paid(self):
+        self.paid_status = self.PAID
+        self.paid_date = timezone.now()
+        self.save()
